Add function-form processor
Also use a backtracking checker function to see whether a second pass
is needed with new information from the first pass. For functions
right now this only checks whether one of the processed children
contains a newline, in which case a function call will put each
argument on a new line in a second pass.

The second reprocesses all children, which is not efficient. It might
be better to just adjust the spacing between the args, on the other
hand the base indentation changes for all args, so that'd require
adjustment as well... that could be done in a final step, however.
I'll leave that puzzle for later.