Add attempt to move postfix comments
Turns out that won't be doable, because some of the other comment
logic already would result in situations where comments would look
like postfix comments, but actually align with the next line. One of
the tests already includes this case:

(some-function ; a comment
               arg1
               arg2)

So this commit will be removed again, but this was an initial, naive
approach to move postfix comments to the left of the previous
expression, if there was one.