Refactor multiline? and num-chunks to use zipper navigation
Adding a node to the right is a hack, but I haven't had a better idea
to iterate through all the nodes in the subtree only. This appears to
be 5-10% slower on a test file I've been using, likely due to the
insertion of the new node all the time.

So I'll revert this anyway, but for documentation purposes, this is an
approach I've tried.