Keep whitespace between discarded forms
Discarded forms are nested, and so far only the top level one will be
aligned inside the parent form. All its children would get their
whitespaces stripped, but for now just keep the whitespace as it is.

This actually has the nice effect of not forcing a realignment of
chained discard forms. Probably still can be improved.