Add better handling of trailing comments
Trailing comments and newline comments make a lot of this much more
complicated. Perhaps the code can be refactored again at some point to
make the various rules and special cases more explicit.

For now there are a bunch of new special case tests, many are edge
cases, and they are handled consistently. The rules are:

1. keep trailing comments on the same line, but reduce the spaces to 0
2. newline comments are handled as before, except that there are some
   special cases in interactions with trailing comments on the
   previous line, they must be ignored
3. if at any time an argument with newline comments above it was
   supposed to align to a previous argument, then an attempt is made
   to look for the last indentation after a newline in the current
   form, and that indentation is used
4. if other formatting rules would turn a newline comment into a
   trailing comment, then that is prevented and the comment is aligned
   as in 3.